Fix GCC 4 warnings in many parts of code.
Most warnings are about unitialized values being used, set but unused values, missing function prototypes and parenthesis suggestions.

This is due to these parts to be pretty old, from GCC 3 days and earlier, while GCC 4 is much more picky.

Needed for #21.

+++ src/emx/src/lib/sys/__read.c
@@ -93,5 +93,5 @@ int __read (int handle, void *buf, size_t cbToRead)
         errno = -rc;
         LIBCLOG_ERROR_RETURN_INT(-1);
     }
-    LIBCLOG_RETURN_INT(cbRead);
+    LIBCLOG_RETURN_INT((int)cbRead);
 }
